PW-7179 - Re-add parent AdyenUiComponentProvider
