+++ Model/Ui/TokenUiComponentProvider.php
@@ -19,22 +19,8 @@ use Magento\Vault\Model\Ui\TokenUiComponentInterface;
 use Magento\Vault\Model\Ui\TokenUiComponentProviderInterface;
 use Magento\Vault\Model\Ui\TokenUiComponentInterfaceFactory;
 
-class TokenUiComponentProvider implements TokenUiComponentProviderInterface
+class TokenUiComponentProvider extends AdyenUiComponentProvider implements TokenUiComponentProviderInterface
 {
-    /**
-     * @param TokenUiComponentInterfaceFactory $componentFactory
-     * @param Data $dataHelper
-     */
-    public function __construct(
-        TokenUiComponentInterfaceFactory $componentFactory,
-        Data $dataHelper,
-        Vault $vaultHelper
-    ) {
-        $this->componentFactory = $componentFactory;
-        $this->dataHelper = $dataHelper;
-        $this->vaultHelper = $vaultHelper;
-    }
-
     /**
      * Get UI component for token
      *
@@ -43,22 +29,6 @@ class TokenUiComponentProvider implements TokenUiComponentProviderInterface
      */
     public function getComponentForToken(PaymentTokenInterface $paymentToken): TokenUiComponentInterface
     {
-        $tokenType = $this->vaultHelper->getAdyenTokenType($paymentToken);
-        $details = json_decode($paymentToken->getTokenDetails() ?: '{}', true);
-        $details['icon'] = $this->dataHelper->getVariantIcon($details['type']);
-        $createdAt = new \DateTime($paymentToken->getCreatedAt());
-        $details['created'] = $createdAt->format('Y-m-d');
-        $details['displayToken'] = $tokenType === Recurring::CARD_ON_FILE || is_null($tokenType);
-
-        return $this->componentFactory->create(
-            [
-                'config' => [
-                    'code' => AdyenCcConfigProvider::CC_VAULT_CODE,
-                    TokenUiComponentProviderInterface::COMPONENT_DETAILS => $details,
-                    TokenUiComponentProviderInterface::COMPONENT_PUBLIC_HASH => $paymentToken->getPublicHash()
-                ],
-                'name' => 'Adyen_Payment/js/view/payment/method-renderer/adyen-vault-method'
-            ]
-        );
+        return $this->getCardComponentForToken($paymentToken);
     }
 }
