+++ Plugin/CustomerFilterVaultTokens.php
@@ -11,6 +11,7 @@
 
 namespace Adyen\Payment\Plugin;
 
+use Adyen\Payment\Helper\Vault;
 use Magento\Vault\Model\CustomerTokenManagement;
 use Adyen\Payment\Helper\Recurring;
 
@@ -18,6 +19,8 @@ class CustomerFilterVaultTokens
 {
     /**
      * Returns filtered list of payment tokens for current customer session
+     * Hide token if it is specifically set to SUBSCRIPTION or UNSCHEDULED_CARD_ON_FILE
+     *
      * @param CustomerTokenManagement $customerTokenManagement
      * @param array $customerSessionTokens
      * @return array
